Add fullscreen image viewer and refactor ChatView into smaller files
- Add tap-to-open fullscreen image viewer with NavigationStack toolbar,
  drag-to-dismiss, and system share button
- Extract message bubble views to MessageBubbleViews.swift (~1025 lines):
  MessageGroupRow, MessageBubble, MediaAttachmentView, ReactionChips,
  ReplyPreviewCard, PikaPromptView, PikaHtmlView, PikaWebView, etc.
- Extract FullscreenImageViewer to its own file (~77 lines)
- Fix EmojiPickerSheet blank screen after selecting emoji
- Reduce long press duration from 0.5s to 0.3s
- Only attach image tap gesture when callback is provided

ChatView.swift: 2349 -> 1353 lines
